Job working system functional
The job system will now allow you to enter a job ID and start a timer or
create a new job if the ID doesn't exist. When you stop the timer, the
time spent is saved into the database.

Next:

Delete specific job.
Allow the user to select jobs, and then delete all others.